requirements.txt
chess
mido
PyYAML